Skip package installation if the package was already installed

+++ boost/install.sh
@@ -7,22 +7,32 @@
 set -e
 set -x
 
-# Install gcc
-GCC_INSTALL_SCRIPT_PATH=${GCC_INSTALL_SCRIPT_PATH:-"GCC_INSTALL_SCRIPT_PATH_IN"}
-[[ $GCC_INSTALL_SCRIPT_PATH == http* ]] && wget ${GCC_INSTALL_SCRIPT_PATH} -O /tmp/gcc-install.sh && GCC_INSTALL_SCRIPT_PATH=/tmp/gcc-install.sh
-source $GCC_INSTALL_SCRIPT_PATH
+# Check for existing installation
+SHARE_PREFIX="/usr/local/share/fem-on-colab"
+BOOST_INSTALLED="$SHARE_PREFIX/boost.installed"
 
-# Install zlib
-apt install -y -qq zlib1g-dev
+if [[ ! -f $BOOST_INSTALLED ]]; then
+    # Install gcc
+    GCC_INSTALL_SCRIPT_PATH=${GCC_INSTALL_SCRIPT_PATH:-"GCC_INSTALL_SCRIPT_PATH_IN"}
+    [[ $GCC_INSTALL_SCRIPT_PATH == http* ]] && GCC_INSTALL_SCRIPT_DOWNLOAD=${GCC_INSTALL_SCRIPT_PATH} && GCC_INSTALL_SCRIPT_PATH=/tmp/gcc-install.sh && [[ ! -f ${GCC_INSTALL_SCRIPT_PATH} ]] && wget ${GCC_INSTALL_SCRIPT_DOWNLOAD} -O ${GCC_INSTALL_SCRIPT_PATH}
+    source $GCC_INSTALL_SCRIPT_PATH
 
-# Download and uncompress library archive
-BOOST_ARCHIVE_PATH=${BOOST_ARCHIVE_PATH:-"BOOST_ARCHIVE_PATH_IN"}
-[[ $BOOST_ARCHIVE_PATH == http* ]] && wget ${BOOST_ARCHIVE_PATH} -O /tmp/boost-install.tar.gz && BOOST_ARCHIVE_PATH=/tmp/boost-install.tar.gz
-if [[ $BOOST_ARCHIVE_PATH != skip ]]; then
-    tar -xzf $BOOST_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
-fi
+    # Install zlib
+    apt install -y -qq zlib1g-dev
 
-# Add symbolic links to the MPI libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
-if [[ $BOOST_ARCHIVE_PATH != skip ]]; then
-    ln -fs /usr/local/lib/libboost*so* /usr/lib
+    # Download and uncompress library archive
+    BOOST_ARCHIVE_PATH=${BOOST_ARCHIVE_PATH:-"BOOST_ARCHIVE_PATH_IN"}
+    [[ $BOOST_ARCHIVE_PATH == http* ]] && BOOST_ARCHIVE_DOWNLOAD=${BOOST_ARCHIVE_PATH} && BOOST_ARCHIVE_PATH=/tmp/boost-install.tar.gz && wget ${BOOST_ARCHIVE_DOWNLOAD} -O ${BOOST_ARCHIVE_PATH}
+    if [[ $BOOST_ARCHIVE_PATH != skip ]]; then
+        tar -xzf $BOOST_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
+    fi
+
+    # Add symbolic links to the MPI libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
+    if [[ $BOOST_ARCHIVE_PATH != skip ]]; then
+        ln -fs /usr/local/lib/libboost*so* /usr/lib
+    fi
+
+    # Mark package as installed
+    mkdir -p $SHARE_PREFIX
+    touch $BOOST_INSTALLED
 fi
