+++ gmsh/install.sh
@@ -7,27 +7,37 @@
 set -e
 set -x
 
-# Install h5py (and its dependencies, most notably gcc and mpi4py)
-H5PY_INSTALL_SCRIPT_PATH=${H5PY_INSTALL_SCRIPT_PATH:-"H5PY_INSTALL_SCRIPT_PATH_IN"}
-[[ $H5PY_INSTALL_SCRIPT_PATH == http* ]] && wget ${H5PY_INSTALL_SCRIPT_PATH} -O /tmp/h5py-install.sh && H5PY_INSTALL_SCRIPT_PATH=/tmp/h5py-install.sh
-source $H5PY_INSTALL_SCRIPT_PATH
+# Check for existing installation
+SHARE_PREFIX="/usr/local/share/fem-on-colab"
+GMSH_INSTALLED="$SHARE_PREFIX/gmsh.installed"
 
-# Install OCC (and its dependencies, most notably gcc)
-OCC_INSTALL_SCRIPT_PATH=${OCC_INSTALL_SCRIPT_PATH:-"OCC_INSTALL_SCRIPT_PATH_IN"}
-[[ $OCC_INSTALL_SCRIPT_PATH == http* ]] && wget ${OCC_INSTALL_SCRIPT_PATH} -O /tmp/occ-install.sh && OCC_INSTALL_SCRIPT_PATH=/tmp/occ-install.sh
-source $OCC_INSTALL_SCRIPT_PATH
+if [[ ! -f $GMSH_INSTALLED ]]; then
+    # Install h5py (and its dependencies, most notably gcc and mpi4py)
+    H5PY_INSTALL_SCRIPT_PATH=${H5PY_INSTALL_SCRIPT_PATH:-"H5PY_INSTALL_SCRIPT_PATH_IN"}
+    [[ $H5PY_INSTALL_SCRIPT_PATH == http* ]] && H5PY_INSTALL_SCRIPT_DOWNLOAD=${H5PY_INSTALL_SCRIPT_PATH} && H5PY_INSTALL_SCRIPT_PATH=/tmp/h5py-install.sh && [[ ! -f ${H5PY_INSTALL_SCRIPT_PATH} ]] && wget ${H5PY_INSTALL_SCRIPT_DOWNLOAD} -O ${H5PY_INSTALL_SCRIPT_PATH}
+    source $H5PY_INSTALL_SCRIPT_PATH
 
-# Download and uncompress library archive
-GMSH_ARCHIVE_PATH=${GMSH_ARCHIVE_PATH:-"GMSH_ARCHIVE_PATH_IN"}
-[[ $GMSH_ARCHIVE_PATH == http* ]] && wget ${GMSH_ARCHIVE_PATH} -O /tmp/gmsh-install.tar.gz && GMSH_ARCHIVE_PATH=/tmp/gmsh-install.tar.gz
-if [[ $GMSH_ARCHIVE_PATH != skip ]]; then
-    tar -xzf $GMSH_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
+    # Install OCC (and its dependencies, most notably gcc)
+    OCC_INSTALL_SCRIPT_PATH=${OCC_INSTALL_SCRIPT_PATH:-"OCC_INSTALL_SCRIPT_PATH_IN"}
+    [[ $OCC_INSTALL_SCRIPT_PATH == http* ]] && OCC_INSTALL_SCRIPT_DOWNLOAD=${OCC_INSTALL_SCRIPT_PATH} && OCC_INSTALL_SCRIPT_PATH=/tmp/occ-install.sh && [[ ! -f ${OCC_INSTALL_SCRIPT_PATH} ]] && wget ${OCC_INSTALL_SCRIPT_DOWNLOAD} -O ${OCC_INSTALL_SCRIPT_PATH}
+    source $OCC_INSTALL_SCRIPT_PATH
+
+    # Download and uncompress library archive
+    GMSH_ARCHIVE_PATH=${GMSH_ARCHIVE_PATH:-"GMSH_ARCHIVE_PATH_IN"}
+    [[ $GMSH_ARCHIVE_PATH == http* ]] && GMSH_ARCHIVE_DOWNLOAD=${GMSH_ARCHIVE_PATH} && GMSH_ARCHIVE_PATH=/tmp/gmsh-install.tar.gz && wget ${GMSH_ARCHIVE_DOWNLOAD} -O ${GMSH_ARCHIVE_PATH}
+    if [[ $GMSH_ARCHIVE_PATH != skip ]]; then
+        tar -xzf $GMSH_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
+    fi
+
+    # Add symbolic links to gmsh libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
+    if [[ $GMSH_ARCHIVE_PATH != skip ]]; then
+        ln -fs /usr/local/lib/libgmsh*.so* /usr/lib
+    fi
+
+    # Install X11 for gmsh
+    apt install -y -qq libfontconfig1 libgl1
+
+    # Mark package as installed
+    mkdir -p $SHARE_PREFIX
+    touch $GMSH_INSTALLED
 fi
-
-# Add symbolic links to gmsh libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
-if [[ $GMSH_ARCHIVE_PATH != skip ]]; then
-    ln -fs /usr/local/lib/libgmsh*.so* /usr/lib
-fi
-
-# Install X11 for gmsh
-apt install -y -qq libfontconfig1 libgl1
