+++ mpi4py/install.sh
@@ -7,22 +7,32 @@
 set -e
 set -x
 
-# Install gcc
-GCC_INSTALL_SCRIPT_PATH=${GCC_INSTALL_SCRIPT_PATH:-"GCC_INSTALL_SCRIPT_PATH_IN"}
-[[ $GCC_INSTALL_SCRIPT_PATH == http* ]] && wget ${GCC_INSTALL_SCRIPT_PATH} -O /tmp/gcc-install.sh && GCC_INSTALL_SCRIPT_PATH=/tmp/gcc-install.sh
-source $GCC_INSTALL_SCRIPT_PATH
+# Check for existing installation
+SHARE_PREFIX="/usr/local/share/fem-on-colab"
+MPI4PY_INSTALLED="$SHARE_PREFIX/mpi4py.installed"
 
-# Download and uncompress library archive
-MPI4PY_ARCHIVE_PATH=${MPI4PY_ARCHIVE_PATH:-"MPI4PY_ARCHIVE_PATH_IN"}
-[[ $MPI4PY_ARCHIVE_PATH == http* ]] && wget ${MPI4PY_ARCHIVE_PATH} -O /tmp/mpi4py-install.tar.gz && MPI4PY_ARCHIVE_PATH=/tmp/mpi4py-install.tar.gz
-if [[ $MPI4PY_ARCHIVE_PATH != skip ]]; then
-    tar -xzf $MPI4PY_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
-fi
+if [[ ! -f $MPI4PY_INSTALLED ]]; then
+    # Install gcc
+    GCC_INSTALL_SCRIPT_PATH=${GCC_INSTALL_SCRIPT_PATH:-"GCC_INSTALL_SCRIPT_PATH_IN"}
+    [[ $GCC_INSTALL_SCRIPT_PATH == http* ]] && GCC_INSTALL_SCRIPT_DOWNLOAD=${GCC_INSTALL_SCRIPT_PATH} && GCC_INSTALL_SCRIPT_PATH=/tmp/gcc-install.sh && [[ ! -f ${GCC_INSTALL_SCRIPT_PATH} ]] && wget ${GCC_INSTALL_SCRIPT_DOWNLOAD} -O ${GCC_INSTALL_SCRIPT_PATH}
+    source $GCC_INSTALL_SCRIPT_PATH
 
-# Add symbolic links to the MPI libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
-if [[ $MPI4PY_ARCHIVE_PATH != skip ]]; then
-    ln -fs /usr/local/lib/libmca*.so* /usr/lib
-    ln -fs /usr/local/lib/libmpi*.so* /usr/lib
-    ln -fs /usr/local/lib/libopen*.so* /usr/lib
-    ln -fs /usr/local/lib/ompi*.so* /usr/lib
+    # Download and uncompress library archive
+    MPI4PY_ARCHIVE_PATH=${MPI4PY_ARCHIVE_PATH:-"MPI4PY_ARCHIVE_PATH_IN"}
+    [[ $MPI4PY_ARCHIVE_PATH == http* ]] && MPI4PY_ARCHIVE_DOWNLOAD=${MPI4PY_ARCHIVE_PATH} && MPI4PY_ARCHIVE_PATH=/tmp/mpi4py-install.tar.gz && wget ${MPI4PY_ARCHIVE_DOWNLOAD} -O ${MPI4PY_ARCHIVE_PATH}
+    if [[ $MPI4PY_ARCHIVE_PATH != skip ]]; then
+        tar -xzf $MPI4PY_ARCHIVE_PATH --strip-components=2 --directory=/usr/local
+    fi
+
+    # Add symbolic links to the MPI libraries in /usr/lib, because Colab does not export /usr/local/lib to LD_LIBRARY_PATH
+    if [[ $MPI4PY_ARCHIVE_PATH != skip ]]; then
+        ln -fs /usr/local/lib/libmca*.so* /usr/lib
+        ln -fs /usr/local/lib/libmpi*.so* /usr/lib
+        ln -fs /usr/local/lib/libopen*.so* /usr/lib
+        ln -fs /usr/local/lib/ompi*.so* /usr/lib
+    fi
+
+    # Mark package as installed
+    mkdir -p $SHARE_PREFIX
+    touch $MPI4PY_INSTALLED
 fi
